refactor: [006-GROUP/CREATE] service intreface 제거, dto 변경, entity 변경으로 인한 패키지명/service명 변경 #19
