+++ src/main/java/com/valuewith/tweaver/groupMember/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.valuewith.tweaver.member.repository;
+package com.valuewith.tweaver.groupMember.repository;
 
-import com.valuewith.tweaver.member.entity.GroupMember;
+import com.valuewith.tweaver.groupMember.entity.GroupMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<GroupMember, Long> {
