+++ src/main/java/com/valuewith/tweaver/groupMember/service/GroupMemberService.java
@@ -1,12 +1,12 @@
-package com.valuewith.tweaver.member.service;
+package com.valuewith.tweaver.groupMember.service;
 
 import com.valuewith.tweaver.chat.entity.ChatRoom;
 import com.valuewith.tweaver.constants.ApprovedStatus;
 import com.valuewith.tweaver.constants.MemberRole;
 import com.valuewith.tweaver.group.entity.TripGroup;
-import com.valuewith.tweaver.member.entity.GroupMember;
-import com.valuewith.tweaver.member.repository.MemberRepository;
-import com.valuewith.tweaver.user.entity.Member;
+import com.valuewith.tweaver.groupMember.entity.GroupMember;
+import com.valuewith.tweaver.groupMember.repository.MemberRepository;
+import com.valuewith.tweaver.menber.entity.Member;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -16,10 +16,9 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Slf4j
 @Transactional
-public class MemberServiceImpl implements MemberService{
+public class GroupMemberService {
   private final MemberRepository memberRepository;
 
-  @Override
   public void setMember(TripGroup tripGroup, Member member, ChatRoom chatRoom) {
     GroupMember groupMember = GroupMember.builder()
         .memberRole(MemberRole.LEADER)
