+++ src/main/java/com/valuewith/tweaver/menber/dto/MemberDto.java
@@ -1,7 +1,7 @@
-package com.valuewith.tweaver.user.dto;
+package com.valuewith.tweaver.menber.dto;
 
 import com.valuewith.tweaver.alert.dto.AlertDto;
-import com.valuewith.tweaver.member.dto.MemberDto;
+import com.valuewith.tweaver.groupMember.dto.GroupMemberDto;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.Data;
@@ -11,8 +11,8 @@ import org.springframework.stereotype.Component;
 @Component
 @Data
 @NoArgsConstructor
-public class UserDto {
-  private Long userId;
+public class MemberDto {
+  private Long memberId;
   private String email;
   private String password;
   private String nickName;
@@ -20,6 +20,6 @@ public class UserDto {
   private String gender;
   private String profileUrl;
   private Boolean isSocial;
-  private List<MemberDto> members = new ArrayList<>();
+  private List<GroupMemberDto> groupMembers = new ArrayList<>();
   private List<AlertDto> alerts = new ArrayList<>();
 }
